+++ src/frontend/apps/desk/src/features/teams/team-management/components/index.ts
@@ -1,3 +1,1 @@
-export * from './CardCreateTeam';
-export * from './TeamInfo';
-export * from './TeamLayout';
+export * from './TeamView';